Thank you for your wonderful work!

**YOLOv10 with End2End (TRT EfficientNMS)**
**Note:** The primary purpose of employing End2End is to utilize ONNX models on TensorRT. If you choose not to use TensorRT, you should proceed with the standard ONNX export process.

I've created a forked repository from the original, adding End-to-End support for ONNX export. The changes can be found in https://github.com/laugh12321/yolov10/commit/fd013f544228e72da47edba8e7e6f63b6c78fc8c .

- **Support for End-to-End ONNX Export:** Added support for end-to-end ONNX export.

- **Configuration Variables:** Use the following flags to configure the model:
   - `nms`: Enabled export End2End with TRT EfficientNMS.

   - `max_det`: The maximum number of detections to output per image.
   - `iou`: The scalar threshold for IOU (additional boxes that have high IOU overlap with previously selected boxes are removed).
   - `conf`: The scalar threshold for score (low scoring boxes are removed).

**Export:**

```bash
$ yolo export model=D:\laugh\Downloads\yolov10s.pt format=onnx opset=13 simplify max_det=100 conf=0.25 iou=0.65 nms
Ultralytics YOLOv8.1.34 🚀 Python-3.12.3 torch-2.3.0+cpu CPU (AMD Ryzen 7 5700X 8-Core Processor)
YOLOv10s summary (fused): 293 layers, 8096880 parameters, 0 gradients

PyTorch: starting from 'D:\laugh\Downloads\yolov10s.pt' with input shape (1, 3, 640, 640) BCHW and output shape(s) ((1, 1), (1, 100, 4), (1, 100), (1, 100)) (31.4 MB)

ONNX: starting export with onnx 1.16.1 opset 13...
[W shape_type_inference.cpp:1972] Warning: The shape inference of TRT::EfficientNMS_TRT type is missing, so it may result in wrong shape inference for the exported graph. Please consider adding it in symbolic function. (function UpdateReliable)
[W shape_type_inference.cpp:1972] Warning: The shape inference of TRT::EfficientNMS_TRT type is missing, so it may result in wrong shape inference for the exported graph. Please consider adding it in symbolic function. (function UpdateReliable)
[W shape_type_inference.cpp:1972] Warning: The shape inference of TRT::EfficientNMS_TRT type is missing, so it may result in wrong shape inference for the exported graph. Please consider adding it in symbolic function. (function UpdateReliable)
[W shape_type_inference.cpp:1972] Warning: The shape inference of TRT::EfficientNMS_TRT type is missing, so it may result in wrong shape inference for the exported graph. Please consider adding it in symbolic function. (function UpdateReliable)
ONNX: simplifying with onnxsim 0.4.36...
ONNX: export success ✅ 1.4s, saved as 'D:\laugh\Downloads\yolov10s.onnx' (27.8 MB)

Export complete (3.4s)
Results saved to D:\laugh\Downloads
Predict:         yolo predict task=detect model=D:\laugh\Downloads\yolov10s.onnx imgsz=640
Validate:        yolo val task=detect model=D:\laugh\Downloads\yolov10s.onnx imgsz=640 data=None
Visualize:       https://netron.app
💡 Learn more at https://docs.ultralytics.com/modes/export
```

**Inference:**

You can utilize the [TensorRT-YOLO](https://github.com/laugh12321/TensorRT-YOLO) repository to efficiently infer TensorRT models equipped with the EfficientNMS plugin. For detailed instructions, please refer to [this link](https://github.com/laugh12321/TensorRT-YOLO/blob/main/demo/detect/README.md#%E6%A8%A1%E5%9E%8B%E6%8E%A8%E7%90%86).