@jameslahm 2024-05-25
Thank you for your contribution! Due to that YOLOv10 does not rely on the NMS, it can be directly exported to end2end onnx without the EfficientNMS plugin, like https://github.com/THU-MIG/yolov10?tab=readme-ov-file#export