@laugh12321 2024-05-25
@jameslahm Thank you for your reply. I noticed that the official end2end ONNX model does not use the EfficientNMS plugin, but instead uses the `v10postprocess` function. This function selects the top `max_det` detection boxes with the highest confidence scores. During prediction, it further filters based on `conf`. rom the results, The effect of EfficientNMS is equivalent to executing these two steps.