@jameslahm 2024-05-25
Thanks. But the EfficientNMS plugin would also check and filter overlapping boxes, which YOLOv10 doesn't need?