@laugh12321 2024-05-25
Thank you for your correction. You are right, the EfficientNMS plugin checks and filters overlapping detection boxes, which is a step not needed in YOLOv10. This is one of the reasons why YOLOv10 can achieve efficient end-to-end deployment without relying on the EfficientNMS plugin.
﻿
Below are the performance test results using `trtexec` on YOLOv10s with `max_det` set to 100, tested on GPU RTX 2080Ti 22GB on AMD Ryzen 7 5700X 8-Core/ 128GB RAM.
﻿
**With plugin:**
```
Latency: min = 2.26025 ms, max = 3.58774 ms, mean = 2.46407 ms, median = 2.42133 ms, percentile(90%) = 2.60785 ms, percentile(95%) = 2.81689 ms, percentile(99%) = 3.05481 ms
```
﻿
**Without plugin:**
```
Latency: min = 2.14551 ms, max = 4.20935 ms, mean = 2.44236 ms, median = 2.34351 ms, percentile(90%) = 2.80798 ms, percentile(95%) = 2.97852 ms, percentile(99%) = 3.24243 ms
```
﻿
This might serve as an alternative option for end-to-end deployment, highlighting the advantages of YOLOv10 compared to other models.